Refactored each scraper into own file
