+++ index.js
@@ -3,204 +3,12 @@ const express = require('express')
 const cors = require('cors');
 const assert = require("assert");
 
+const scrapers = require("./scrapers");
+
 const port = process.env.JDANKS_PORT || 80;
 const app = express()
 
 const idToData = new Map();
-
-async function youtube(id) {
-    const {data} = await axios.get(`https://youtube.com/channel/${id}`);
-
-    const avatar = data.match(/channelMetadataRenderer.*?avatar.*?thumbnails.*?url".*?"(.*?)"/s)[1];
-    const name = data.match(/channelMetadataRenderer.*?title".*?"(.*?)"/s)[1];
-
-    const {data: livedata} = await axios.get(`https://youtube.com/channel/${id}/live`);
-    let live, title;
-    try {
-        title = livedata.match(/videoPrimaryInfoRenderer".*?"title".*?"runs".*?"text".*?"(.*?)"/s)[1];
-        live = livedata.match(/playabilityStatus.*?status".*?"(.*?)"/s)[1] !== "LIVE_STREAM_OFFLINE";
-    } catch (e) {
-        live = false;
-    }
-
-    let r = {
-        avatar,
-        name,
-        live,
-        id,
-        platform: 'youtube',
-    }
-
-    if (live) {
-        r.title = title;
-        let viewers = livedata.match(/videoPrimaryInfoRenderer".*?"viewCount".*?"runs".*?"text".*?"(.*?)"/s)[1]?.split(' ')[0];
-        r.viewers = Number(viewers.replace(/[.,]/g, ""));
-    }
-
-    return r;
-}
-
-async function dlive(id) {
-    const { data } = await axios.post('https://graphigo.prd.dlive.tv/',
-        `{"query":"query{userByDisplayName(displayname: \\"${id}\\") {livestream{ view title } avatar}}"}`);
-
-    const response = data.data.userByDisplayName;
-    const live = !!response.livestream;
-
-    let r = {
-        platform: "dlive",
-        id, name: id,
-        avatar: response.avatar,
-        live,
-    }
-
-    if(live) {
-        r.viewers = response.livestream.view;
-        r.title = response.livestream.title;
-    }
-
-    return r;
-}
-
-async function bitwave(id) {
-    const {data: _data} = await axios.get(`https://api.bitwave.tv/v1/channels/${id}`);
-    if(!_data?.success) return {};
-    const data = _data.data;
-    return {
-        live: data.live,
-        name: data.name,
-        avatar: data.avatar,
-        id,
-        platform: "bitwave",
-        viewers: data.viewCount,
-        title: data.title,
-    };
-}
-
-async function robotstreamer(id, name) {
-    const {data: _data} = await axios.get(`http://api.robotstreamer.com:8080/v1/get_robot/${id}`);
-    const data = _data[0];
-    const live = data.status === "online";
-    const title = data.robot_name;
-    const viewers = data.viewers;
-    return {
-        live, name, id,
-        platform: "robotstreamer",
-        title, viewers
-    }
-}
-
-const trovo_id_memoization = new Map();
-async function trovo(username) {
-    const token = process.env.TROVO_CLIENT_ID;
-
-    assert( !!token );
-
-    if( !trovo_id_memoization.has(username) ) {
-        const {data} = await axios.post(
-            'https://open-api.trovo.live/openplatform/getusers',
-            {user: [username]},
-            {
-                headers: {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'Client-ID': token,
-                }
-            }
-        );
-        trovo_id_memoization.set(username, data.users[0].channel_id );
-    }
-
-    assert(trovo_id_memoization.has(username));
-
-    const {data} = await axios.post(
-        'https://open-api.trovo.live/openplatform/channels/id',
-        {channel_id: trovo_id_memoization.get(username)},
-        {
-            headers: {
-                'Content-Type': 'application/json',
-                'Accept': 'application/json',
-                'Client-ID': token,
-            }
-        }
-    );
-
-    return {
-        live: data.is_live,
-        title: data.live_title,
-        viewers: data.current_viewers,
-        id: username, platform: "trovo",
-        name: username,
-        avatar: data.profile_pic,
-    }
-}
-
-const twitch_avatars = new Map();
-let twitch_access_token;
-
-async function twitch_init() {
-    const {data} = await axios.post(
-        `https://id.twitch.tv/oauth2/token`
-        + `?client_id=${process.env.TWITCH_CLIENT_ID}`
-        + `&client_secret=${process.env.TWITCH_CLIENT_SECRET}`
-        + `&grant_type=client_credentials`
-    );
-    twitch_access_token = data;
-
-    setInterval(async () => {
-        twitch_access_token = (await axios.post(
-            `https://id.twitch.tv/oauth2/token`
-            + `?client_id=${process.env.TWITCH_CLIENT_ID}`
-            + `&client_secret=${process.env.TWITCH_CLIENT_SECRET}`
-            + `&grant_type=client_credentials`
-        )).data;
-    }, twitch_access_token.expires_in);
-}
-
-async function twitch(username) {
-    const access_token = (await twitch_access_token).access_token;
-
-    // Update avatar every fifth scrape of the same username
-    // or; update every avatars every 25 minutes
-    if( !twitch_avatars.has(username) || twitch_avatars.get(username)[1] === 5 ) {
-        const {data: user_data} = await axios.get(`https://api.twitch.tv/helix/users?login=${username}`, {
-            headers: {
-                'Client-Id': process.env.TWITCH_CLIENT_ID,
-                'Authorization': `Bearer ${access_token}`
-            }
-        });
-        twitch_avatars.set(username, [user_data.data[0].profile_image_url, 5]);
-    } else {
-        twitch_avatars.get(username)[1]--;
-    }
-
-    const {data: _data} = await axios.get(`https://api.twitch.tv/helix/streams?user_login=${username}`, {
-        headers: {
-            'Client-Id': process.env.TWITCH_CLIENT_ID,
-            'Authorization': `Bearer ${access_token}`
-        }
-    });
-    const stream_data = _data.data[0];
-
-    return {
-        live: stream_data.type === "live",
-        name: username,
-        id: username,
-        avatar: twitch_avatars.get(username)[0],
-        platform: "twitch",
-        title: stream_data.title,
-        viewers: stream_data.viewer_count
-    };
-}
-
-const scrapers = new Map([
-    ["youtube", youtube],
-    ["dlive", dlive],
-    ["bitwave", bitwave],
-    ["robotstreamer", robotstreamer],
-    ["trovo", trovo],
-    ["twitch", twitch],
-]);
 
 const rateLimit = require("express-rate-limit");
 const limiter = rateLimit({
@@ -239,14 +47,12 @@ async function scrape(platform, id, name) {
 const loadPeople = (async (people) => {
     console.info("Populating scrape data...");
 
-    await twitch_init();
-
     // multithread all initial scrapes, wait for them all to finish
     await Promise.all(people.map(async (person, i) => {
         const platform = person[0];
         const id = person[1];
 
-        scrape(platform, id, person[2]);
+        await scrape(platform, id, person[2]);
         console.info(`Scraped ${id}!`);
 
         setTimeout(() => {
